Game module events rename; Validation about delegate move and deduplicate;
