+++ proto/g4alchain/game/events.proto
@@ -3,14 +3,14 @@ package g4alentertainment.g4alchain.game;
 
 option go_package = "github.com/G4AL-Entertainment/g4al-chain/x/game/types";
 
-message EventCreate {
+message EventCreateProject {
     string symbol       = 1;
     string name         = 2;
     string description  = 3;
     string creator      = 4;
 }
 
-message EventUpdate {
+message EventUpdateProject {
     string symbol       = 1;
     string name         = 2;
     string description  = 3;
