+++ x/assetfactory/keeper/validation.go
@@ -28,58 +28,26 @@ func (k Keeper) ValidateArgsClass(symbol string, description string, name string
 	return nil
 }
 
-func (k Keeper) ValidateProjectOwnershipOrDelegateByProject(ctx sdk.Context, creator string, symbol string) error {
-	// Checking project existing and related to this game developer or delegate
-	project, found := k.gameKeeper.GetProject(ctx, symbol)
-	if !found {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "project invalid symbol (%s)", symbol)
-	}
-
-	// Check if msg.Creator included in valFound.Delegate
-	isDelegate := false
-	for _, del := range project.Delegate {
-		if del == creator {
-			isDelegate = true
-			break
-		}
-	}
-	// Checks if the msg creator is the same as the current owner
-	if creator != project.Creator && !isDelegate {
-		return sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "incorrect owner nor delegate address")
-	}
-	return nil
-}
-
 func (k Keeper) ValidateProjectOwnershipOrDelegateByClassId(ctx sdk.Context, creator string, symbol string) error {
 	// Checking if symbol/classID exists via x/nft
 	classFound, found := k.nftKeeper.GetClass(ctx, symbol)
 	if !found {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "symbol of classID not found x/nft (%s)", symbol)
 	}
-
 	// check on map to what project is associated with
 	classMapFound, found := k.GetClass(ctx, classFound.Symbol)
 	if !found {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "symbol of classID not found x/assetfactory (%s)", symbol)
 	}
-
 	// Checking project existing and related to this game developer or delegate
 	project, found := k.gameKeeper.GetProject(ctx, classMapFound.Project)
 	if !found {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "project invalid symbol (%s)", classMapFound.Project)
 	}
-
-	// Check if msg.Creator included in valFound.Delegate
-	isDelegate := false
-	for _, del := range project.Delegate {
-		if del == creator {
-			isDelegate = true
-			break
-		}
-	}
-	// Checks if the msg creator is the same as the current owner
-	if creator != project.Creator && !isDelegate {
-		return sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "incorrect owner nor delegate address")
+	// Check delegate
+	err := k.gameKeeper.ValidateDelegate(creator, project)
+	if err != nil {
+		return err
 	}
 	return nil
 }
