+++ x/assetfactory/types/expected_keepers.go
@@ -17,6 +17,8 @@ type PermissionKeeper interface {
 type GameKeeper interface {
 	// Methods imported from game should be defined here
 	GetProject(ctx sdk.Context, symbol string) (val gametypes.Project, found bool)
+	ValidateDelegate(creator string, project gametypes.Project) error
+	ValidateProjectOwnershipOrDelegateByProject(ctx sdk.Context, creator string, symbol string) error
 }
 
 type NftKeeper interface {
